OneCloud/20220112-江皓哲-imageColorization/requirements.txt
Flask==2.0.2
numpy==1.16.5
Pillow==9.0.0
scikit_image==0.15.0